refactor(e2e): one nvidia-smi XML schema for every assertion
The encoder/FBC checks (#636) and the JPEG/OFA and process checks each decoded
`nvidia-smi -q -x` into their own types. Fold them into a single schema in
nvidiasmi_xml.go with one decode entry point and one scalar reader, so an
assertion that needs another element adds a field rather than a parser.

Readings stay strings across the board because nvidia-smi renders "N/A" in the
same element that otherwise carries a number, and the leading-integer read
already covered "1500 us", so "35 %" and "6000 MiB" need no separate handling.
nvidiasmi_temperature.go keeps only the text-table parsing, which is there
because the defect it guards is which row labels nvidia-smi prints at all.
